+++ modules/bluetooth.nix
@@ -1,0 +1,6 @@
+{
+  nixosModule = { pkgs, ... }: {
+    hardware.bluetooth.enable = true;
+    services.blueman.enable = true;
+  };
+}